+++ convex/share.test.ts
@@ -24,45 +24,3 @@ test('sharing a chat works if there is a snapshot', async () => {
 });
 
 // TODO: Test that cloning messages does not leak a more recent snapshot or later messages
-
-test('unreferenced snapshots are deleted', async () => {
-  const t = setupTest();
-  const { sessionId, chatId } = await createChat(t);
-  const storageId1 = await t.run((ctx) => ctx.storage.store(new Blob(['Hello, world!'])));
-  await t.mutation(internal.snapshot.saveSnapshot, { sessionId, chatId, storageId: storageId1 });
-  const storageId2 = await t.run((ctx) => ctx.storage.store(new Blob(['foobar'])));
-  await t.mutation(internal.snapshot.saveSnapshot, { sessionId, chatId, storageId: storageId2 });
-  // `storageId1` should be deleted because it was overwritten by `storageId2`
-  await expect(t.run((ctx) => ctx.storage.get(storageId1))).resolves.toBeNull();
-  // `storageId2` should not be deleted
-  const blob = await t.run(async (ctx) => {
-    const blob = await ctx.storage.get(storageId2);
-    return blob?.text();
-  });
-  expect(blob).toBe('foobar');
-});
-
-test('referenced snapshots are not deleted', async () => {
-  const t = setupTest();
-  const { sessionId, chatId } = await createChat(t);
-  const storageId1 = await t.run((ctx) => ctx.storage.store(new Blob(['Hello, world!'])));
-  await t.mutation(internal.snapshot.saveSnapshot, { sessionId, chatId, storageId: storageId1 });
-
-  await t.mutation(api.share.create, { sessionId, id: chatId });
-  const storageId2 = await t.run((ctx) => ctx.storage.store(new Blob(['foobar'])));
-  await t.mutation(internal.snapshot.saveSnapshot, { sessionId, chatId, storageId: storageId2 });
-
-  // `storageId1` should not be deleted because it is referenced by the share
-  const blob1 = await t.run(async (ctx) => {
-    const blob = await ctx.storage.get(storageId1);
-    return blob?.text();
-  });
-  expect(blob1).toBe('Hello, world!');
-
-  // `storageId2` should not be deleted
-  const blob2 = await t.run(async (ctx) => {
-    const blob = await ctx.storage.get(storageId2);
-    return blob?.text();
-  });
-  expect(blob2).toBe('foobar');
-});
